refactor 1st step - split handelr
