+++ verifier/pkg/verifier/loader/ruleTable_test.go
@@ -216,3 +216,5 @@ func TestMatchWithRequest(t *testing.T) {
 
 	}
 }
+
+// TestMatchNamespace
